State managers are not global anymore
