+++ src/Layers/xrRender/R_Backend.h
@@ -21,6 +21,7 @@
 #include "Layers/xrRenderPC_R4/r_backend_lod.h"
 #include "Layers/xrRenderDX11/StateManager/dx11StateManager.h"
 #include "Layers/xrRenderDX11/StateManager/dx11ShaderResourceStateCache.h"
+#include "Layers/xrRenderDX11/StateManager/dx11StateCache.h"
 #endif
 
 #include "FVF.h"
@@ -580,6 +581,10 @@ public:
     dx11StateManager StateManager;
     dx11SamplerStateCache SSManager;
     dx11ShaderResourceStateCache SRVSManager;
+
+    dx11StateCache<ID3DRasterizerState, D3D_RASTERIZER_DESC> RSManager;
+    dx11StateCache<ID3DDepthStencilState, D3D_DEPTH_STENCIL_DESC> DSSManager;
+    dx11StateCache<ID3DBlendState, D3D_BLEND_DESC> BSManager;
 #endif // USE_DX11
 };
 #pragma warning(pop)
